Manage first call after webhook creation and avoid none github call
I noticed that github make a call when you create a webhook. Previously
Phubme was trying to handle it and was crashing because the comment was
not present inside the body. I now check the content of the body and
process only if it's a valid json.
It also check if the call are made from github with a simple validation
on header.