Fix first github call when creating webhook (#16)
* Manage first call after webhook creation and avoid none github call

I noticed that github make a call when you create a webhook. Previously
Phubme was trying to handle it and was crashing because the comment was
not present inside the body. I now check the content of the body and
process only if it's a valid json.
It also check if the call are made from github with a simple validation
on header.

* Use pattern matching for valid_github_payload rather than cond

* Simplify how Plug get port

* Separate how we process github payload + style changes